+++ backend/vulkan/src/commonMain/kotlin/io/technoirlab/rhi/vulkan/VulkanBuffer.kt
@@ -9,11 +9,11 @@ import io.technoirlab.vulkan.DeviceMemory as VkDeviceMemory
 internal abstract class VulkanBuffer(
     internal val buffer: VkBuffer,
     private val memory: VkDeviceMemory,
-    override val size: UInt
+    override val size: ULong
 ) : Buffer {
 
-    override fun updateData(source: Source, size: UInt, offset: UInt) = memScoped {
-        memory.copyData(source, size.toULong(), offset.toULong())
+    override fun updateData(source: Source, size: ULong, offset: ULong) = memScoped {
+        memory.copyData(source, size, offset)
     }
 
     override fun close() {
